Three files of code
Put all three in your working directory.
Open run-model.R and run it (it calls in the other two).
This simulates the data, sets up everything for the MCMC and runs it.
The output is then plotted in various ways.
I hope my comments explain the structure clearly.
I'd welcome any feedback on how to structure this better. Also, I need to include the code to test the sub-grid explicitly, but will upload that separately.